:truck: refactor: RestResponse Move Domain to DTO folder
